Changed version to 1.1.0-SNAPSHOT
Would be cool try to release it as SNAPSHOT first and do some testing.
If you don't use "-SNAPSHOT" suffix, the artifact will be published
in the Sonatype Staging Repository.

To understand how staging works, and how to move artifact from
'staging' to 'releases', check this documentation "Release It":
https://docs.sonatype.org/display/Repository/Sonatype+OSS+Maven+Repository+Usage+Guide#SonatypeOSSMavenRepositoryUsageGuide-8.ReleaseIt

You just need to add the PGP plugin and create a key, before being able
to publish on releases.