Add BoundedRange type class
The existing `Range` type class is inflexible, only allowing
ascending ranges that are left-inclusive and right-exclusive.

This introduces a bounded range type class that more closely
models the mathematical concept of a range, allowing for open
and closed range boundaries, as well as both ascending and
descending ranges.

This new type class, `BoundedRange`, witnesses an `Hlist` of
`Nat` numbers given a starting and ending bound.

Bounds are either `Inclusive` or `Exclusive`, which includes
or excludes the `Nat` at that end of the range.

Ascending ranges start with a number less than the ending
number, and descending ranges start with a number greater
than the ending number.

Fancy symbolic syntax is also available with type aliases.

Example usage:
```scala
import nat._; import syntax.nat.range._

the[_1 *--* _4].apply() == _1 :: _2 :: _3 :: _4 :: HNil
the[_1 :--* _4].apply() == _2 :: _3 :: _4 :: HNil
the[_1 *--: _4].apply() == _1 :: _2 :: _3 :: HNil
the[_1 :--: _4].apply() == _2 :: _3 :: HNil
the[_4 *--* _1].apply() == _4 :: _3 :: _2 :: _1 :: HNil
```