Add compileTime macro to measure compilation times
`compileTime` is a simple macro that measures the compilation
time of a code fragment. Like `illTyped` it takes Scala code as
string, measures the time it takes to parse and typecheck it and
returns that time as a `Duration`.

Example usage:
```scala
scala> test.compileTime("Generic[(Int, Option[String])]")
res0: scala.concurrent.duration.FiniteDuration = 34911863 nanoseconds
```