Added `UnzipFunctions` typeclass
This typeclass witnesses that an `HList` type `L` relates two other
`HList` types `Args` and `Returns`, where each element in `L` is of the
form `A => B`, where `A` is the corresponding element from `Args`, and
`B` is the corresponding element from `Returns`. In some ways it
complements `ZipApply`, but to use `ZipApply`, you must already know the
type of `Args`. `UnzipFunctions` allows inferring the `Args` and
`Returns`, either with just `L`, or with `L` and one of `Args` and
`Returns`, or with all three.