Remove `UnzipFunctions`
I did this because in the course of writing up an example for
`ZipApply`, I realized that `UnzipFunctions` is unnecessary. Its only
use case that isn't covered by `ZipApply` is inferring the result type
of function application without inferring argument types. But there's no
reason that the argument types can't be inferred, and it hurts nothing
to let them be inferred even if you don't need them.