Added type class to support folding a Coproduct by handling each type in the order given by the Coproduct type signature.
For instance, a `Coproduct` `c` of type `Int :+: String :+: Boolean :+: CNil`
could be folded into a `Double` as follows:

val result = c.foldCases[Double]
               .atCase(i => math.sqrt(i))
               .atCase(s => s.length.toDouble)
	       .atCase(b => if (b) 100.0 else -1.0)

This provides some benefits over existing methods for folding over
`Coproduct`s.  Unlike the `Folder` type class, this one does not require a
polymorphic function with a stable identifier, so the syntax is somewhat
lightweight and better suited to situations where the folding function is not
reused (e.g., parser combinator libraries).

Additionally, unlike directly folding over a `Coproduct` with pattern matching
over `Inl` and `Inr` injectors, this type class guarantees that the resulting
fold is exhaustive.  It is also possible to partially fold a `Coproduct` (as
long as cases are handled in the order specified by the `Coproduct` type
signature), which makes it possible to incrementally fold a `Coproduct`.