Fix type loss with `Comapped` and a few other typeclasses
Similarly to https://github.com/milessabin/shapeless/pull/621, the
implicit recursive case of `Comapped` lacks an explicit type argument
for its tail's `Out`. This hurts type inference in some cases. This PR
adds the type argument in order to fix this problem.

Although `Comapped` was my immediate problem, I noticed several other
typeclasses that appear to have the same problem for which this
relatively mechanical fix can be trivially applied. I included a few
more of them in this PR, but there are still several others (in general,
wherever a recursive implicit case refers to a type member of its
implicit argument, type inference can be improved by eliminating that
reference and lifting it to a type argument to the implicit derivation).
I'm happy to make an effort to apply this change more universally, or
users can just continue to make PRs as their need arises. Let me know
what you think @milessabin.