Try to fix more unsoundness in Typeable (#579)
Specifically, prior to this patch, with a type such as:

trait T { type Inner }
class K[P] extends T { type Inner = T; case class Foo(t:T) }

the default derived Typeable for K[P]#Foo would see Foo as a simple
unparameterized case class, and perform casting with a simple
isInstanceOf check.

With this patch, after all other checks for special derived
Typeables are made, before we fall back to simpleTypeable,
we check to see if any values have a type that looks different
from the inside as from the outside, which indicates that the
erased type will not contain information about that type. At
that point, the situation is the same as for a parameterized
type; we can safely derive a Typeable for a product, but not
for an arbitrary type.