Support for Annotation[Option[A], T]
Currently, `Annotation` just aborts compilation when type `T` is not
annotated with `A`. Although it's possible to allow non-annotated case
by implicit priorities, It would be better to use `Option[A]` to avoid
duplication in complex generic derivations.

This PR additionaly allows `Annotation[Option[A], T]` by returning
`Option[A]` in runtime, keeping current behavior of `Annotation[A ,T]`
the same.