Avoid use of mutable state in macro classloader
Recent versions of Scala have an experimental option
(`-Ycache-macro-classloader`) to reuse a macro classloader across
multiple (possibly concurrent) compilations. It is motivated by
performance to reduce overhead of classloading and of the macro
implementation running without time to JIT.

However, this requires macro implenentations to avoid races and leaks
with any classloader-global state.

This commit introduces a generic facility for managing macro
state that should have the same lifecycle as a `Global#Run` and
uses it in the `Cached` macro.