Replace CtorDtor with simplified FromTo
FromTo decouples and simplifies the strategy to generate the
from and to functions. This will let us more easily add
other strategies in the future and refine the choice of which
strategy to use in which case.

This is a prerequisite for fixing #1059 but does not yet fix it.
The issue demonstrates that in certain cases we might want to use
fromApply and toGetters for case classes.